requirements.txt
paho-mqtt>=1.6.0,<2.0.0
typing-extensions>=4.0.0; python_version<"3.10"
